# CodeEdit: Improve render time by 2x