… scene/gui/text_edit.cpp