with gdscript syntax highlighting:
before: 30.0385588566243ms avg draw time (1089 samples)
after: 13.0933055755083ms avg draw time (1089 samples)

without syntax highlighting:
before: 17.9521260292186ms avg draw time (1089 samples)
after: 12.9270402524754ms avg draw time (1089 samples)

#### many improvements to do with:

avoiding redraw:
- use `set_process_internal(true)` instead of queue_redraw to avoid drawing 2-3 times in the same frame when `gui_input()` is called
  - tests needed to be updated bc of this
- `set_code_hint` and `set_code_hint_draw_below` avoid redrawing when the values dont change

improving render speed:
- cache line syntax highlighting
- line numbers:
  - caching line number shaped text
  - simple enough to use TextServer RIDs directy instead of `TextLine`
  - only render line numbers on screen
- `_main_gutter_draw_callback`: use `get_hovered_gutter()` instead of multiple calls to the more expensive `get_local_mouse_pos()`
- cache `total_visible_line_count` and `Line::line_count`
- delay calculating text max width and height in case its not needed
- `GutterInfo`, which contains a custom callable in both of godots script editor gutters, was getting copied once per gutter per line per frame,  

some less impactful (more code style) things ..
- avoid creating HashMaps every draw frame, just use vectors
- const-ify copies of Vectors to avoid possible copy-on-write when theres no write

its obvious that most of the problem was the line syntax highlighting being called every frame
but whats not obvious is that searching through the dictionary and converting to and from variants was a very significant part of it

the performance improvements are especially noticeable on debug builds, as there arent any 'double draws' making the editor lag when scrolling

benchmark: [CodeEditText.zip](https://github.com/user-attachments/files/15737660/CodeEditText.zip)
tested with godot compiled using: `production=yes optimize=size tests=yes deprecated=no scu_build=yes`
bc it uses the GDScriptSyntaxHighlighter, to benchmark you have open the project in the editor