@KoBeWi 2024-06-07
> use set_process_internal(true) instead of queue_redraw to avoid drawing 2-3 times in the same frame

Drawing is already limited to once per frame. It's called `queue_redraw()`, because it queues one redraw no matter how many times you call it. A different result would be a bug.