@AThousandShips 2024-06-07
How does `queue_redraw` get called multiple times per frame though? The queue gets flushed a few times but it's in different phases of the frame, and unless some code calls `queue_redraw` in different stages it won't be drawn multiple times?