@rune-scape 2024-06-07
it does get called in multiple stages, specifically in `gui_input` and in `NOTIFICATION_INTERNAL_PHYSICS_PROCESS` and somehow in another place , sometimes, (beware i tested this on 4.2) i tested how many times per iteration it was drawing, and it was between 2 and 3 every frame when scrolling vigarously