@rune-scape 2024-06-07
i see.. i.. assumed it was simply copied over from `VScroll` and never changed, im curious what the bug is, i havent noticed it yet