@rune-scape 2024-06-08
sorry for not being clear about how it draws multiple times every frame but i didn't expect it to be unbelievable
i could split out out the redraw part of this PR, it seems controversial, but heres a test to show im not making this up:
[CodeEditMultipleDrawTest.zip](https://github.com/user-attachments/files/15746201/CodeEditMultipleDrawTest.zip)

i tested it just now on master on a debug linux build and it seems to draw 8! times per process iteration when scrolling vigarously

and @MewPurPur, i asked about the smooth scrolling bug bc in this PR ive changed the TextEdit to use `NOTIFICATION_INTERNAL_PROCESS` instead of physics process (it seemed like you were saying that causes a bug), and i have some people using this, i'd like to know how serious the bug is.