@MewPurPur 2024-06-08
https://github.com/godotengine/godot/issues/28385