@lawnjelly 2024-06-08
> it very much does draw multiple times per process loop (maybe different meaning than frame) bc queue_redraw adds a message to the MessageQueue which gets flushed multiple times per iteration

Some options off top of my head:
* Frame specific `MessageQueue`.
* Bespoke list that is flushed once per frame (i.e. `CanvasItem` redraw queue).

I've used the latter multiple times, notably for the physics interpolation. An advantage here is you'd have tighter control of the timing of the calls.

Frame specific `MessageQueue` I suspect would inevitably end up getting flushed multiple times and you'd end up with the same problem.

Another potential snag in either case is if the update caused side effects, like queuing the update of another `CanvasItem`, or adding stuff to the `MessageQueue`, or even setting up recursive loops, so you might need to limit what could be done in these updates, or deal with this possibility explicitly.