@MewPurPur 2024-06-10
I'm finding a lot of instances where _draw() is called twice in 4.3 when it was only called once in 4.2.2, could that be related? I haven't set on to make a MRP yet, but someone else encountered the same:

> _draw runs for me twice too for some reason
> but only after nesting the script with the _draw in another control
> for me it was panel