@rune-scape 2024-06-14
> However, I get a reproducible slowdown in terms of editor FPS when opening the [Platformer 2D](https://github.com/godotengine/godot-demo-projects/tree/master/2d/platformer)'s player.gd script in the editor

> this may point out at some scalability issues in the new code (e.g. methods that have O(N²) scaling has replaced methods that were previously O(N)).

@Calinou im quite sure theres only one place this happened (replacing a hashmap lookup with `Vector.has`), and it only applies to having multiple cursors, 

when i tested this i didnt get any meaningful slowdown, (probly bc i dont have a 4k monitor)
but i have a slight suspicion that changing it to use `NOTIFICATION_INTERNAL_PROCESS` caused this, and now that its changed back to use `NOTIFICATION_INTERNAL_PHYSICS_PROCESS` it may not have that slowdown