@rune-scape 2024-08-12
@kitbdev thank you for reviewing!!! everything u mentioned should be fixed now,

seems the max line height/width being invalidated by setting it to -1 was not a good idea,, lol