@arkology 2024-09-06
I have tested speed improvement by changing zoom inside script editor and difference is really noticeable🚀